pyproject.toml
[project]
name = "neuralese"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "UFO-101", email = "josephmiller101@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "datasets==3.2",
    "einops>=0.8.0",
    "fairseq2==0.3.0rc1",
    "jaxtyping>=0.2.36",
    "numpy",
    "pillow>=11.1.0",
    "sonar-space>=0.3.2",
    "torch==2.5.1",
    "transformer-lens==2.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "beartype>=0.14.1",
    "ipykernel>=6.29.5",
    "ipython>=8.31.0",
    "pre-commit>=4.0.1",
    "pyright>=1.1.391",
    "pytest>=8.3.4",
    "pytest-beartype>=0.1.1",
    "tqdm>=4.67.1",
    "wandb>=0.19.2",
]

[tool.ruff]
# Advised for jaxtyping (see https://docs.kidger.site/jaxtyping/faq/#flake8-or-ruff-are-throwing-an-error)
# For single dimension tensors, you have to prepend a space to the type annotation
lint.ignore = ["F722"]
line-length = 88

[tool.pytest.ini_options]
# Only typecheck during tests because it messes with debugger
beartype_packages = "neuralese"

[tool.pyright]
reportMissingParameterType = true
reportUnknownParameterType = true

[[tool.uv.index]]
url = "https://fair.pkg.atmeta.com/fairseq2/whl/rc/pt2.5.1/cu124"
